+++ src/utils/store_util.py
@@ -3,9 +3,10 @@ import os
 
 import pandas as pd
 
-from src import portfolio, collection_store, battle_store
+from src import portfolio, collection_store, battle_store, season_balances_info, season_battle_info
 from src.api import spl
 from src.configuration import store, config
+from src.static.static_values_enum import Format
 from src.utils import progress_util
 
 
@@ -42,7 +43,7 @@ def load_stores():
     for store_name in get_store_names():
         store_file = get_store_file(store_name)
         if os.path.isfile(store_file):
-            #TODO investigate the low_memory
+            # TODO investigate the low_memory
             # DtypeWarning: Columns (6,14) have mixed types. Specify dtype option on import or set low_memory=False.
             #   store.__dict__[store_name] = pd.read_csv(store_file, index_col=0)
             store.__dict__[store_name] = pd.read_csv(store_file, index_col=0, low_memory=False)
@@ -138,7 +139,7 @@ def get_seasons_played_list():
         temp_end_dates.end_date = pd.to_datetime(temp_end_dates.end_date)
 
         last_id = temp_end_dates.loc[(temp_end_dates.end_date > first_date)].id.min()
-        return temp_end_dates.sort_values('id', ascending=False).loc[(temp_end_dates.id >= last_id-1)].id.to_list()
+        return temp_end_dates.sort_values('id', ascending=False).loc[(temp_end_dates.id >= last_id - 1)].id.to_list()
     else:
         return list()
 
@@ -155,7 +156,15 @@ def get_last_season_values(df, users, season_id_column='season_id'):
     return df.loc[(df.player.isin(users)) & (df[season_id_column] == df[season_id_column].max())].copy()
 
 
-def update_data():
+def is_last_season_processed(account, current_season_data):
+    if not (store.season_sps.empty or store.season_sps.loc[store.season_sps.player == account].empty):
+        last_season = store.season_sps.loc[store.season_sps.player == account].season_id.max()
+        if last_season == current_season_data['id'] - 1:
+            return True
+    return False
+
+
+def update_battle_log():
     progress_util.set_daily_title('Update collection')
     collection_store.update_collection()
 
@@ -167,3 +176,45 @@ def update_data():
 
     save_stores()
     progress_util.update_daily_msg('Done')
+
+
+def update_season_log():
+    progress_util.set_season_title("Season update process initiated")
+    update_season_end_dates()
+    current_season_data = spl.get_current_season()
+
+    for account in get_account_names():
+        progress_util.update_season_msg('Start season update for: ' + str(account))
+        if not is_last_season_processed(account, current_season_data):
+            if spl.is_season_reward_claimed(account, current_season_data):
+                season_balances_info.update_balances_store(account, current_season_data)
+                store.season_modern_battle_info = season_battle_info.get_season_battles(account,
+                                                                                        store.season_modern_battle_info.copy(),
+                                                                                        Format.modern,
+                                                                                        current_season_data)
+                store.season_wild_battle_info = season_battle_info.get_season_battles(account,
+                                                                                      store.season_wild_battle_info.copy(),
+                                                                                      Format.wild,
+                                                                                      current_season_data)
+        else:
+            progress_util.update_season_msg("No seasons to process for: " + str(account))
+
+    save_stores()
+    progress_util.set_season_title("Season update done")
+    progress_util.update_season_msg('Done')
+
+
+def update_data(battle_update=True, season_update=False):
+    try:
+        if not spl.is_maintenance_mode():
+            if battle_update:
+                update_battle_log()
+
+            if season_update:
+                update_season_log()
+        else:
+            logging.info("Splinterlands server is in maintenance mode skip this update cycle")
+    except Exception as e:
+        logging.error("Exception during update data")
+        logging.exception(e)
+