+++ CHANGES.md
@@ -1,6 +1,13 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
+## 0.8.0
+- FIX: When splinterlands is in maintenance mode skip update, causing application crashes
+- Update readme for executable + change linux to ubuntu in github workflow
+- Change server mode wait time. After processing daily and season reschedule next pull cycle to 30 minutes.
+- Add season pull information to server mode. Including some feedback when rewards are not claimed yet.
+- Update page title and remove the "Updating..." 
+
 ## 0.7.0
 - FIX: Home page default from date filter causing error when not loaded fully.
 - FIX: Deposit/withdraw exception without account